Tests/LibWeb/Text/expected/css/system-color-highlight-color-scheme.txt

Light HighlightText: rgb(255, 255, 255)
Dark HighlightText: rgb(20, 20, 20)
Dark CanvasText: rgb(235, 235, 235)
